UI & Level Design
Added UI Control
Added Player Selection
Added Menu Navigation
Added Game Menu
Added Level
Added moving Texture
Added more 3D Models

+++ Assets/ProCore/ProBuilder/Shader/pb_FacePicker.shader
@@ -1,6 +1,6 @@
 ﻿// Upgrade NOTE: replaced 'mul(UNITY_MATRIX_MVP,*)' with 'UnityObjectToClipPos(*)'
 
-Shader "Hidden/ProBuilder/FacePicker"
+Shader "Hidden/ProBuilder/FacePicker" 
 {
 	Properties {}
 
@@ -43,7 +43,7 @@ Shader "Hidden/ProBuilder/FacePicker"
 				return o;
 			}
 
-			float4 frag (v2f i) : COLOR
+			half4 frag (v2f i) : COLOR
 			{
 				return i.color;
 			}
